+++ predictive_punter/__init__.py
@@ -1,5 +1,7 @@
 __version__ = '1.0.0a1'
 
 
+from . import runner
+
 from .command import Command
 from .scrape import ScrapeCommand