+++ predictive_punter/__init__.py
@@ -1,6 +1,7 @@
 __version__ = '1.0.0a1'
 
 
+from . import race
 from . import runner
 
 from .command import Command