+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Extend racing_data.Runner objects with a calculate_expected_times method (from @justjasongreen)
+- Extend racing_data.Runner objects with a races_per_year property (from @justjasongreen)
 
 ### Changed
 - Log the item associated with an exception in process_collection (from @justjasongreen)