+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Extend racing_data.Race objects with a total_value property (from @justjasongreen)
 - Extend racing_data.Runner objects with a calculate_expected_times method (from @justjasongreen)
 - Extend racing_data.Runner objects with a races_per_year property (from @justjasongreen)
+- Add the Sample class (from @justjasongreen)
 
 ### Changed
 - Log the item associated with an exception in process_collection (from @justjasongreen)